+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+### Changes
+* Lowered Cryo Aspect freeze duration
+
 ## 1.26.5
 ### Changes
 * Enchantments are no longer consumed in creative mode