+++ changelog.md
@@ -7,6 +7,7 @@
   * Only one level now
   * Steals 25 ticks
   * Ticks stolen now scale with attack speed and no longer with weapon % attack damage
+* Slightly increased Trident enchantability (7 -> 8)
 
 ### Fixes
 * Player attack cooldown not calculated correctly in hammer knockback