+++ changelog.md
@@ -8,6 +8,7 @@
   * Steals 25 ticks
   * Ticks stolen now scale with attack speed and no longer with weapon % attack damage
 * Slightly increased Trident enchantability (7 -> 8)
+* Vanilla and ISO Shields enchantability changes no longer require Shields+ data pack enabled 
 
 ### Fixes
 * Player attack cooldown not calculated correctly in hammer knockback